added simple view for single post
-- removed single post component (no extra functionality justifying new component)
-- fixed post identification type from string to long

+++ src/main/java/com/rakayby/blog/controller/PostController.java
@@ -44,7 +44,7 @@ public class PostController {
     }
 
     @GetMapping(ApiEndPoints.PostController.GET_POST)
-    public ResponseEntity<?> getById(@RequestParam(required = true) Integer id) {
+    public ResponseEntity<?> getById(@RequestParam(required = true) Long id) {
         final Optional post = this.postFacade.getById(id);
         if (post.isPresent()) {
             return ResponseEntity.ok(post.get());
